% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwargs.R
\name{args_and_kwargs}
\alias{args_and_kwargs}
\title{Does Python-esque \code{*args, **kwargs} argument separation}
\usage{
args_and_kwargs(...)
}
\arguments{
\item{\dots}{Whatever mix of named and unnamed arguments you want}
}
\description{
This function will return a named a separated list of named and unnamed arguments
as quosures.
}
\examples{
# Like how I made `share_discrete_scales`

share_discrete_scales <- function(gg_obj, ...) {
  akw <- args_and_kwargs(...)
  # unnamed arguments are ggplot scale functions
  geom_func_list <- purrr::map(akw$args, rlang::eval_tidy)
  # named args are to be passed into these functions
  geoms <- purrr::map(geom_func_list, ~quo_to_args(., akw$kwargs))
  answer <- purrr::reduce(geoms, ggplot2::`\%+\%`, .init=gg_obj)
  return(answer)
}
}

