% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolabel_latex_captions.R
\name{autolabel_latex_figs}
\alias{autolabel_latex_figs}
\title{Autolabel figures in Latex}
\usage{
autolabel_latex_figs()
}
\description{
I hate having to add \verb{\\\\label\{fig:blah\}} to \code{fig.cap}
in Latex .Rmd files.  There's probably a better way to do
this, but if you call this function, it will automatically
add \verb{\\\\label\{fig:<label>\}} to the \code{fig.cap} option of any
chunk with a \code{fig.cap} and a label, substituting whitespace
stretches (after trimming it)
}
