% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolabel_latex_captions.R
\name{beep_on_knit_error}
\alias{beep_on_knit_error}
\title{Beep on knit error}
\usage{
beep_on_knit_error(beep_sound = 9, sleep = 2)
}
\arguments{
\item{beep_sound}{Input to \code{beepr::beep()} for what sound to play}

\item{sleep}{The amount of seconds to sleep before moving on (so that the sound isn't cut off after it knits)}
}
\description{
This function requires the \code{beepr} package. It will first set the global chunk function to \code{error=TRUE} (so knitr won't stop on errors), but then adds a hook that will catch these errors, play a sound, and then throw an error anyway, shutting it down. \cr \cr
Will not beep for non-code chunk errors.
}
