% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_df.R
\name{drop_empty_subs}
\alias{drop_empty_subs}
\title{Drop rows with empty nested data frames}
\usage{
drop_empty_subs(.data, data_col_name, empty_df = TRUE)
}
\arguments{
\item{.data}{A data frame / tibble}

\item{data_col_name}{The column name of the nested data frames, bare or as a string.}

\item{empty_df}{If \code{TRUE}, also drops rows with data frames that have 0 rows.}
}
\description{
Drops rows that for a given column of nested data frames, have \code{NULL} values. Optionally, rows with nested data frames that have 0 rows can also be dropped.
}
