% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbars.R
\name{errorbars}
\alias{errorbars}
\title{Make errorbars with one function}
\usage{
errorbars(gg_obj, ...)
}
\arguments{
\item{gg_obj}{the ggplot object}

\item{\dots}{values for the stat_summary}
}
\description{
Pipe the ggplot object into this. If not immediately after
\code{ggplot(\dots)}, use \code{\%+\%} instead of \code{+} to add layers until
the ggplot object is piped into the function.
}
\examples{
ggplot(aes(x = x, y = y)) \%+\%
   geom_point() \%>\%
   errorbars() +
   xlab("WHAAAT")
}

