% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbars_new.R
\name{geom_dots_and_bars}
\alias{geom_dots_and_bars}
\title{Make dotplots with errorbars with a single function}
\usage{
geom_dots_and_bars(..., binaxis = "y", stackdir = "center", dotsize = 0.5,
  alpha = 0.2)
}
\arguments{
\item{\dots}{Additional arguments for the \code{\link[ggplot2]{stat_summary}} pair.}
}
\value{
A list of the \code{stat_summary} objects
}
\description{
Adds a dotplot, 95\% CI errorbars, and a point representing the mean
to a ggplot object. Almost identical to \code{\link{errorbars}}, but
with the obvious additional \code{\link[ggplot2]{geom_dotplot}}. \cr
Right now, only uses default values for errorbars
}
\examples{
df <- data.frame(
  y<-rnorm(100),
  x<-factor(c(rep("Label1",50), rep("Label2",50)))
)

ggplot(df, aes(x = x, y = y)) +
  geom_dots_and_bars(color="red",fill="red") +
  xlab("WHAAAT")
}
\seealso{
\code{\link{errorbars}}
}

