% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbars_new.R
\name{geom_errorbars}
\alias{geom_errorbars}
\title{Make ggplot errorbars and mean with a single function}
\usage{
geom_errorbars(...)
}
\arguments{
\item{\dots}{Additional arguments for the \code{\link[ggplot2]{stat_summary}} pair.}
}
\value{
A list of the \code{stat_summary} objects
}
\description{
Often, automatically bootstrapped errorbars in \code{ggplot2} are accompanied
by a dot indicating the mean of the errorbars. Generally, this is done with
two calls of \code{\link[ggplot2]{stat_summary}}, one for the bars and one
for the dot.  This function is just a shortcut that will add both.
}
\examples{
df <- data.frame(
  y<-rnorm(100),
  x<-factor(c(rep("Label1",50), rep("Label2",50)))
)

ggplot(df, aes(x = x, y = y)) +
  geom_point(alpha=0.3) +
  geom_errorbars(color="red") +
  xlab("WHAAAT")
}
\seealso{
\code{\link{geom_dots_and_bars}}
}
