% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment_plus.R
\name{geom_segment_plus}
\alias{geom_segment_plus}
\title{Extended geom_segment}
\usage{
geom_segment_plus(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  arrow = NULL,
  arrow.fill = NULL,
  lineend = "butt",
  linejoin = "round",
  shorten.limit = c("warn", "drop", "ignore"),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, \dots, stat, parse, inherit.aes}{See the documentation for \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}}
}
\description{
Still a WIP. The dropping and warning of segments switching directions isn't really working correctly yet.  \code{geom_segment_plus} is an updated version of mo-seph's Stack Overflow code (https://stackoverflow.com/q/14647988/4560765), which no longer works. mo-seph's code in turn is an extension of \code{ggplot2}'s \code{geom_segment}, but with the ability to offset segments that overlap, and to shorten the end or beginning of the segment.
}
\details{
This code basically has three new possible parameters:
\itemize{
\item \strong{\code{shorten.start}/\code{shorten.end}}: this moves the start/end of the segment closer to the midpoint by a proportion
\item \strong{\code{shorten.start.abs}/\code{shorten.end.abs}}: this moves the start/end of the segment closer to the midpoint by an absolute value
\item \strong{\code{offset}}: if two segments fully overlap, this offsets them slightly
}
}
\examples{
d <- data.frame(x=c(-1,1),
           y=c(-1,1),
           x.to=c(2,-1),
           y.to=c(-1,1))
d \%>\%
  ggplot(aes( x=x, y=y ) ) +
  geom_point() +
  geom_segment_plus(aes( x=x, y=y, xend=x.to, yend=y.to ),
                    lineend="round", arrow=arrow(length=unit(0.15, "inches")),
                    offset=0.00, shorten.start=0.5, shorten.end=0,
                    shorten.end.abs=0.15,
                    shorten.limit = 'drop')
d \%>\%
  ggplot(aes( x=x, y=y ) ) +
  geom_point() +
  geom_segment_plus(aes( x=x, y=y, xend=x.to, yend=y.to ),
                    lineend="round", arrow=arrow(length=unit(0.15, "inches")),
                    offset=0.00, shorten.start=1, shorten.end=0,
                    shorten.end.abs=1.25,
                    shorten.limit = 'drop')
}
