% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Get ggplot default colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{the number of different colors in the palette}
}
\description{
Taken from John Colby's answer here: https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette, this function lets you get the default palette from ggplot2.
}
