% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dput_reminder.R
\name{object_to_input}
\alias{object_to_input}
\title{How to turn a object into a text representation/input}
\description{
This isn't a function---it's just a note to document something that I found to be very useful. \cr
\cr
Just use \code{dput(x)}. E.g., if \code{x} is \code{c(1,2,3)}, then \code{dput(x)} is "c(1, 2, 3)".  Thats it!
}
