% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolabel_latex_captions.R
\name{set_blog_knitr_opts}
\alias{set_blog_knitr_opts}
\title{Setting knitr options for my blog}
\usage{
set_blog_knitr_opts(version = NULL)
}
\description{
This is my lame attempting at standardizing default knitr options for my blog, with some versioning for consistency / replicability. \code{get_blog_knitr_opts()} is the getter version of this function.
}
