% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolabel_latex_captions.R
\name{set_widget_hooks}
\alias{set_widget_hooks}
\title{Set knitr hooks for htmlwidgets with Jekyll/GitHub Pages}
\usage{
set_widget_hooks(
  dep_dir = knitr::opts_chunk$get("plotly.savepath"),
  base_path = knitr::opts_chunk$get("proj.basedir"),
  hrefFilter = function(x) paste0("/", x)
)
}
\arguments{
\item{dep_dir}{The directory you want to save the dependencies to.}

\item{base_path}{The directory you want to make the dependency links relative to.
For example, if your post is at the url \verb{yoursite.com/this_post/}, then you'll
want to make \code{base_path} the home directory of your project.}

\item{hrefFilter}{This function lets you perform any additional manipulations
to the dependency links. For some reason, when I give my base directory as my
project's home directory, the links come back without the necessary "/" in front
of them. This function adds that to each link.}
}
\description{
Modified from my gist: https://gist.github.com/burchill/8392b2a753652e24a35a8a1dd707c1b1.
This functions sets a \code{knitr} hook so that any HTML widgets that were
printed (i.e., objects that inherit the \code{'htmlwidget'}, like those from
the \code{htmlwidgets} or \code{plotly} packages) will work with a Jekyll system,
like the one used for GitHub Pages.
}
\details{
It essentially sets a hook so that, when the document is finished being
knitted, it moves all the dependencies necessary for the widgets to a
directory, and then adds HTML code to the document to load those files
from their new location. Additionally, it sets the default
\code{screenshot.force} chunk option to \code{FALSE}, so \code{knitr} doesn't try to
use a screenshot instead of the widget.

See https://www.zachburchill.ml/plotly_with_jekyll/ for background
and (https://gist.github.com/burchill/9df0f6245ea7768e5b6bbd0a1c22db08)
for the old, bad version of this script.
}
