% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart_r.R
\name{start_fresh}
\alias{start_fresh}
\title{A 'safer' command-line restart (for RStudio)}
\usage{
start_fresh(counter = 0)
}
\arguments{
\item{counter}{Do not specify unless you set it to \code{2}, which will force a restart without prompting the user.}
}
\value{
\code{NULL} (because I'm a lazy programmer)
}
\description{
Removes all currently loaded objects and restarts RStudio, after prompting
the user if they really want to do it. \emph{Do NOT change/specify the \code{counter}
argument unless you set it to \code{2}, which forces a restart without prompting}.
}
\details{
When I want to restart my RStudio R session anew, I have, in the past,
clicked on the broom icon, clicked 'yes' to remove all the current objects
in the environment, clicked on 'Session', and then clicked 'Restart R'.
This function does all that with one command, and prompts the user to confirm,
as a way of being safe.

This code takes inspiration and credit from two sources: Adam Lee Perelman's answer at https://stackoverflow.com/questions/6313079/quit-and-restart-a-clean-r-session-from-within-r, and Romain Francois' \code{nothing} package (https://github.com/romainfrancois/nothing/blob/master/R/zzz.R).
}
