% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{stat_dots_and_bar}
\alias{stat_dots_and_bar}
\alias{geom_dots_and_bars}
\title{Make dotplots with error bars with a single function}
\usage{
stat_dots_and_bar(..., binaxis = "y", stackdir = "center",
  dotsize = 0.5, alpha = 0.2)

geom_dots_and_bars(..., binaxis = "y", stackdir = "center",
  dotsize = 0.5, alpha = 0.2)
}
\arguments{
\item{\dots}{Additional arguments for the \code{\link[ggplot2]{stat_summary}} pair.}

\item{binaxis}{Set by default to "y"}

\item{stackdir}{Set by default to "center"}

\item{dotsize}{Set by default to 0.5}

\item{alpha}{Set by default to 0.2}
}
\value{
A list of the \code{stat_summary} objects
}
\description{
Adds a dotplot, 95\% CI error bars, and a point representing the mean
to a ggplot object. Almost identical to \code{\link{stat_errorbar}}, but
with the obvious additional \code{\link[ggplot2]{geom_dotplot}}. \cr
\cr
Right now, only uses default values for the error bars.
}
\details{
\code{geom_dots_and_bars} is the deprecated name for \code{stat_dots_and_bar}. I realized that
\code{ggplot2} already had a \code{\link[ggplot2]{geom_errorbar}}, which would be very
confusing for \code{geom_dots_and_bars}'s sister function, \code{\link{geom_errorbars}}.
To keep the naming similarity, and because \code{geom_dots_and_bars} operated more
like a \code{stat} function (e.g., \code{stat_summary}) than a \code{geom}, it made
more sense to change its name.
}
\examples{
df <- data.frame(
  y<-rnorm(100),
  x<-factor(c(rep("Label1",50), rep("Label2",50)))
)

ggplot(df, aes(x = x, y = y)) +
  stat_dots_and_bar(color="red",fill="red") +
  xlab("WHAAAT")
}
\seealso{
\code{\link{stat_errorbar}}
}
