% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{stat_errorbar}
\alias{stat_errorbar}
\alias{geom_errorbars}
\title{Make ggplot error bars and mean with a single function}
\usage{
stat_errorbar(...)

geom_errorbars(...)
}
\arguments{
\item{\dots}{Additional arguments for the \code{\link[ggplot2]{stat_summary}} pair.}
}
\value{
A list of the \code{stat_summary} objects
}
\description{
Often, automatically bootstrapped error bars in \code{ggplot2} are accompanied
by a dot indicating the mean of the errorbars. Generally, this is done with
two calls of \code{\link[ggplot2]{stat_summary}}, one for the bars and one
for the dot.  This function is just a shortcut that will add both.
}
\details{
\code{geom_errorbars} is the deprecated name for \code{stat_errorbar}. I realized that
\code{ggplot2} already had a \code{\link[ggplot2]{geom_errorbar}}, which would be very
confusing. Additionally, since \code{geom_errorbars} operated more like a \code{stat} function
(e.g., \code{stat_summary}), it made more sense to change its name.
}
\examples{
df <- data.frame(
  y<-rnorm(100),
  x<-factor(c(rep("Label1",50), rep("Label2",50)))
)

ggplot(df, aes(x = x, y = y)) +
  geom_point(alpha=0.3) +
  stat_errorbar(color="red") +
  xlab("WHAAAT")
}
\seealso{
\code{\link{stat_dots_and_bar}}
}
