% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_abs_text.R
\docType{data}
\name{stat_moments}
\alias{stat_moments}
\alias{MomentLabel}
\title{Displaying skew/kurtosis text in plots}
\usage{
stat_moments(mapping = NULL, data = NULL, ..., geom = c("abs_text",
  "abs_label"), moment = c("skewness", "kurtosis", "both"),
  sig = FALSE, excess_kurtosis = FALSE, digits = 1,
  alternative = c("less", "greater", "two.sided"), inherit.aes = TRUE,
  parse = FALSE)
}
\arguments{
\item{mapping, data, \dots, inherit.aes, parse}{See \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} for details.}

\item{geom}{Determines which geom to use, \code{\link[=geom_abs_text]{geom_abs_text()}} or \code{\link[=geom_abs_label]{geom_abs_label()}}. By default, it uses \code{geom_abs_text()}.}

\item{moment}{A string determining which moment to display. Can be one of three values: \code{"skewness"}, \code{"kurtosis"}, or \code{"both"}, which displays both moments.}

\item{sig}{A logical; if true, will test the skewness for significance using \code{\link[moments:agostino.test]{moments::agostino.test()}}, i.e., the D'Agostino test of skewness. Significance will be indicated via asterisks.}

\item{excess_kurtosis}{A logical; if \code{TRUE}, kurtosis will be expressed as "excess" kurtosis (i.e., kurtosis - 3, as 3 is the kurtosis of a normal distribution). If kurtosis is not displayed, this will be ignored.}

\item{digits}{The number of digits after the decimal place to display for the moment values.}

\item{alternative}{A string specifying the alternative hypothesis for the D'Agostino test. Must be one of \code{"less"} (default) \code{"two.sided"} or \code{"greater"}. You can specify just the initial letter. If \code{sig} = \code{FALSE}, this will be ignored.}
}
\description{
\code{stat_moments()} summarises the data supplied to the x-axis, and draws text that displays the skewness and/or kurtosis of the data, with a variety of options. This is almost chiefly meant to be used in conjunction with a density plot, such as \code{\link[ggplot2:geom_density]{ggplot2::geom_density()}} or \code{\link[ggplot2:stat_density]{ggplot2::stat_density()}}.  Since this object is returning text, it needs to be given coordinates on where to be placed. It requires the aesthetics \code{xpos} and \code{ypos} (see \code{\link[=geom_abs_text]{geom_abs_text()}}), which are coordinates (from 0-1) relative to the panel/facet panel the text is to be displayed in.
}
\examples{
make_log_normal <- function(n, mu, sd, name) {
  log_mu <- log(mu)
  df <- data.frame(x=exp(rnorm(n, log_mu, sd=sd)))
  df$Name <- name
  df
}

new_df <- rbind(make_log_normal(1000, 100, 1, "Distr1"),
                make_log_normal(1000, 500, 0.3, "Distr2"),
                make_log_normal(1000, 900, 0.5, "Distr3"))

ggplot(new_df, aes(x=x, color=Name)) +
  geom_density() +
  facet_wrap(~Name, scales="free") +
  stat_moments(aes(xpos=0.5, ypos=0.75),
               sig = TRUE,
               moment = "both",
               fontface="bold") +
  theme_bw()

}
\keyword{datasets}
