% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolabel_latex_captions.R
\name{stop_floating}
\alias{stop_floating}
\title{Stop figures from floating in LaTeX}
\usage{
stop_floating()
}
\description{
When I knit R Markdown files to LaTeX, I end up putting \code{\\FloatBarrier} after every figure
so they don't float past where I want them in the document. Using \code{knitr}'s hooks, I can make this
process happen automatically. \cr
\cr
In order to do that, you first need to include the LaTeX package \code{placeins}
(i.e., via \code{header-includes: \\usepackage{placeins}}).
}
