% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_pass.R
\name{zead}
\alias{zead}
\title{Print the head of a data frame/tibble}
\usage{
zead(df, n = 6L, ...)
}
\arguments{
\item{df}{A data frame, tibble, etc.}

\item{n}{Number of rows to grab with \code{head}}

\item{\dots}{Additional arguments for \code{print.data.frame}}
}
\description{
This is just a convenience function that takes the head of a data frame and prints it using \code{print.data.frame()}, which will display all olumns, even if the data frame is a tibble.
}
