% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifelse-esque.R
\name{zifelse}
\alias{zifelse}
\title{Nested `ifelse` statements with less typing}
\usage{
zifelse(...)
}
\arguments{
\item{\dots}{Paired unnamed arguments, where the first in each pair is an expression that evaluates to a logical vector, and the second is the replacement value. The last argument needs to be a single, unpaired default replacement value.}
}
\value{
A vector of the same length as the logical vector that is the first argument
}
\description{
\emph{USE \code{dplyr}'s \code{\link[dplyr]{case_when}} for actual code!} This function is just code for my (Zach Burchill's) personal
reference in the future. I anticipate that the framework that I wrote here for might be useful for me in the future.
}
\examples{

a <- runif(1000)
zifelse(a >= 0   & a <= 0.33, 1,
        a > 0.33 & a <= 0.66, 2,
        3)
}

