% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zummarise.R
\name{zummarise}
\alias{zummarise}
\title{Summarises and then ungroups a tbl}
\usage{
zummarise(.data, ...)
}
\arguments{
\item{.data}{A tbl. I believe that all main verbs are S3 generics and provide methods for \code{\link[dplyr]{tbl_df}}, \code{\link[dplyr]{tbl_dt}}, and \code{\link[dplyr]{tbl_sql}} since this is built on \code{dplyr} code.}

\item{...}{From \code{dplyr}: Name-value pairs of summary functions. The name will be the
name of the variable in the result. The value should be an expression
that returns a single value like \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.

These arguments are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.}
}
\value{
An object of the same class as \code{.data}. \strong{All} grouping levels will be dropped. \cr \cr
Data frame row names are silently dropped. To preserve, convert to an explicit variable.
}
\description{
When I first started using \code{dplyr}, I thought \code{\link[dplyr]{summarise}}
would completely ungroup the object it was called on. Later, I learned that
it only peeled back the last-named grouping argument. To make sure my code was
doing what I wanted it to, I created this function as a way of implementing what
I thought \code{\link[dplyr]{summarise}} actually did
}
\examples{
zummarise(dplyr::group_by(mtcars, cyl, gear), mean(disp))
}
