% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwargs.R
\name{args_and_kwargs}
\alias{args_and_kwargs}
\title{Separate \dots into Python-esque \code{*args} and \code{**kwargs}}
\usage{
args_and_kwargs(...)
}
\arguments{
\item{\dots}{Whatever mix of named and unnamed arguments you want}
}
\description{
This function will return a named a separated list of named and unnamed arguments
as quosures.
}
\examples{
# Like how I made `share_scales`

share_scales <- function(...) {
  akw <- args_and_kwargs(...)
  # Unnamed arguments are ggplot scales
  geom_func_list <- purrr::map(akw$args, rlang::eval_tidy)
  # Named arguments are to be passed into those scales
  geoms <- purrr::map(geom_func_list, ~quo_to_args(., akw$kwargs))
  return(geoms)
}
}

