% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbars.R
\name{dots_and_bars}
\alias{dots_and_bars}
\title{Make dotplots with errorbars with one function}
\usage{
dots_and_bars(gg_obj, binaxis = "y", stackdir = "center", dotsize = 0.5,
  alpha = 0.2, ...)
}
\arguments{
\item{the}{ggplot object}

\item{\dots}{values for the stat_summary}
}
\description{
Pipe the ggplot object into this, if not immediately, uses %+%
Right now, only uses default values for errorbars
}
\examples{
ggplot(aes(x=x,y=y)) \%+\%
   geom_point() \%>\%
   dots_and_bars() +
   xlab("WHAAAT")
}

