% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbars.R
\name{errorbars}
\alias{errorbars}
\title{Make ggplot errorbars and mean with a single function}
\usage{
errorbars(gg_obj, ...)
}
\arguments{
\item{gg_obj}{The ggplot object (i.e., what is built up from \code{ggplot(\dots)}). If \code{errorbars()} does not immediately follow \code{ggplot(\dots)}, use \code{\%+\%} instead of \code{+} to add the intervening ggplot layers until the ggplot object is piped into the function.}

\item{\dots}{Additional arguments for the \code{\link[ggplot2]{stat_summary}} pair.}
}
\description{
Often, automatically bootstrapped errorbars in \code{ggplot2} are accompanied
by a dot indicating the mean of the errorbars. Generally, this is done with
two calls of \code{\link[ggplot2]{stat_summary}}, one for the bars and one
for the dot.  This function is just a shortcut that will add both.
}
\examples{
ggplot(aes(x = x, y = y)) \%+\%
   geom_point() \%>\%
   errorbars() +
   xlab("WHAAAT")
}

