% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_abs_text.R
\name{geom_abs_text}
\alias{geom_abs_text}
\title{Text positioned relative to a ggplot panel/facet}
\usage{
geom_abs_text(mapping = NULL, data = NULL, stat = "identity", ...,
  parse = FALSE, inherit.aes = TRUE)
}
\arguments{
\item{mapping, data, stat, parse, inherit.aes}{See the documentation for \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}}
}
\description{
\code{geom_abs_text} functions similarly to ggplot2's \code{geom_text}, except that instead of taking \code{x} and \code{y} aesthetics, it requires \code{xpos} and \code{ypos} aesthetics, values from 0-1 that determine the absolute \code{x} and \code{y} coordinates \emph{with respect to whatever viewport they are being plotted in}, be that the full panel or a facet panel. A value of 0.5 indicates the center of the axis, and a value of 1 indicates the far edge of the axis.
\cr \cr This code and its documentation is based off code in the ggplot2 package, and thus is subject to the copy-left licenses of the original package.
}
\details{
This function has only been tested in a few scenarios, and only in \code{ggplot2 v2.2.1}.
}
\examples{
df <- data.frame(
  x = c(99, 0, 1),
  y = c(-100, 0, 100),
  label = c("A","B", "C")
)
ggplot(df, aes(x=x,y=y,label=label, color=label)) +
  geom_point() +
  geom_text() +
  facet_wrap(~label, scales="free") +
  geom_abs_text(aes(xpos=0.5, ypos=0.75, label = paste0("relative: ", label)))

}
