% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zummarise.R
\name{zummarise}
\alias{zummarise}
\title{Summarises and then ungroups df}
\usage{
zummarise(.data, ...)
}
\arguments{
\item{df}{The data frame}
}
\value{
whatever
}
\description{
Because I thought this was how it always worked
}

